START: registration and authentication module in project, finish config spring security for check user role, start validate form

+++ src/main/java/vadim/volin/services/impl/UserServiceImpl.java
@@ -1,32 +1,42 @@
 package vadim.volin.services.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
+import vadim.volin.model.Role;
 import vadim.volin.model.User;
+import vadim.volin.repository.RoleRepository;
 import vadim.volin.repository.UserRepository;
 import vadim.volin.services.UserService;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Service
+@Transactional
 public class UserServiceImpl implements UserService {
 
     @Autowired
     UserRepository userRepository;
 
-//    @Autowired
-//    private BCryptPasswordEncoder bCryptPasswordEncoder;
+    @Autowired
+    RoleRepository roleRepository;
 
     @Override
     public User addUser(User user) {
-//        user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+        if (userRepository.finByMail(user.getUsermail()) != null) {
+            return null;
+        }
+//        List<Role> roleSet = roleRepository.findAll();
+//        user.setRoles(roleSet);
+        user.setRoles("ROLE_USER");
         return userRepository.saveAndFlush(user);
     }
 
     @Override
     public void delete(int id) {
-        userRepository.deleteById(id);
+        if (userRepository.findById(id).isPresent()) {
+            userRepository.deleteById(id);
+        }
     }
 
     @Override
@@ -48,4 +58,5 @@ public class UserServiceImpl implements UserService {
     public List<User> getAllUser() {
         return userRepository.findAll();
     }
+
 }
